requirements.txt
lxml==5.1.0
#lxml
tweepy==4.14.0
#kaleido==0.2.1.post1
kaleido==0.2.1
tldextract==5.1.1
matplotlib==3.8.3
BeautifulSoup4==4.12.3
requests[socks]==2.31.0
playwright==1.41.2
pandas==2.2.1
pillow==10.2.0
#dotenv
python-dotenv==1.0.1
Mastodon.py==1.8.1
playwright_stealth==1.0.5

pika==1.3.0
protobuf==3.20.0
pycountry==23.12.11
